requirements.txt
python-dotenv
discord.py-message-components
openai